+++ scripts/sql/delta/v210-v220/apolloportaldb-v210-v220.sql
@@ -1,0 +1,31 @@
+--
+-- Copyright 2022 Apollo Authors
+--
+-- Licensed under the Apache License, Version 2.0 (the "License");
+-- you may not use this file except in compliance with the License.
+-- You may obtain a copy of the License at
+--
+-- http://www.apache.org/licenses/LICENSE-2.0
+--
+-- Unless required by applicable law or agreed to in writing, software
+-- distributed under the License is distributed on an "AS IS" BASIS,
+-- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+-- See the License for the specific language governing permissions and
+-- limitations under the License.
+--
+# delta schema to upgrade apollo portal db from v2.1.0 to v2.2.0
+
+Use ApolloPortalDB;
+
+ALTER TABLE `App`
+    MODIFY COLUMN `AppId` VARCHAR(64) NOT NULL DEFAULT 'default' COMMENT 'AppID';
+
+ALTER TABLE `Consumer`
+    MODIFY COLUMN `AppId` VARCHAR(64) NOT NULL DEFAULT 'default' COMMENT 'AppID';
+
+ALTER TABLE `Favorite`
+    MODIFY COLUMN `AppId` VARCHAR(64) NOT NULL DEFAULT 'default' COMMENT 'AppID';
+
+ALTER TABLE `Favorite`
+    DROP INDEX `AppId`,
+    ADD INDEX `AppId` (`AppId`);