+++ CHANGES.md
@@ -5,7 +5,7 @@ Release Notes.
 Apollo 2.2.0
 
 ------------------
-
+* [Fix the problem of inconsistent length of appId column](https://github.com/apolloconfig/apollo/pull/4725)
 
 ------------------
 All issues and pull requests are [here](https://github.com/apolloconfig/apollo/milestone/13?closed=1)